% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ci.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Bootstrap confidence interval}
\usage{
bootstrap_ci(
  df,
  f,
  outmod = NULL,
  trtmod = NULL,
  x = NULL,
  y = NULL,
  id = NULL
)
}
\arguments{
\item{df}{A data.frame containing all variables specified in the
provided formula.}

\item{f}{The g-estimation method}

\item{outmod}{The outcome model}

\item{trtmod}{The exposure model}

\item{x}{The exposure / treatment variable}

\item{y}{The outcome variable}

\item{id}{The unique identifier variable of subjects.}
}
\value{
A tibble containing bootstrap confidence intervals of
risk difference, risk ratio and odds ratio
}
\description{
A function to calculate 95\% confidence interval based on bootstrap
method.
}
