% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{dr}
\alias{dr}
\title{Doubly robust estimator used to estimate effect}
\usage{
dr(
  data,
  exposure = NULL,
  y = NULL,
  exposure_model = NULL,
  outcome_model = NULL,
  id = NULL
)
}
\arguments{
\item{data}{A data.frame containing all variables specified in the provided
\code{outcome_model} and \code{exposure_model}.}

\item{exposure}{The exposure (i.e. treatment) variable.}

\item{y}{The outcome variable.}

\item{exposure_model}{The exposure model.}

\item{outcome_model}{The outcome model.}

\item{id}{The unique identifier of subjects.}
}
\value{
A tibble containing risks for exposure level 0 and 1, Risk Difference,
Risk ratio, Odds Ratio
}
\description{
Doubly robust estimator used to estimate effect
}
