% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_modelling_standardization.R
\name{exposure_modeling_standardization}
\alias{exposure_modeling_standardization}
\title{A function to calculate effect of a binary exposure (i.e. treatment) on the
outcome using standardization in exposure-modeling approach.}
\usage{
exposure_modeling_standardization(
  data,
  exposure = NULL,
  y = NULL,
  exposure_model = NULL,
  outcome_model = NULL,
  id = NULL
)
}
\arguments{
\item{data}{A data.frame containing all variables specified in the provided
\code{outcome_model}.}

\item{y}{The outcome variable.}

\item{exposure_model}{The exposure model.}
}
\value{
A tibble containing risks for exposure level 0 and 1, Risk Difference,
Risk ratio, Odds Ratio
}
\description{
A function to calculate effect of a binary exposure (i.e. treatment) on the
outcome using standardization in exposure-modeling approach.
}
