% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_modelling_standardization.R
\name{outcome_modeling_standardization}
\alias{outcome_modeling_standardization}
\title{A function to calculate effect of a binary exposure (i.e. treatment) on the
outcome using standardization in outcome-modeling approach.}
\usage{
outcome_modeling_standardization(
  data,
  exposure = NULL,
  y = NULL,
  exposure_model = NULL,
  outcome_model = NULL,
  id = NULL
)
}
\arguments{
\item{data}{A data.frame containing all variables specified in the provided
\code{outcome_model}.}

\item{exposure}{The exposure (i.e. treatment) variable.}

\item{outcome_model}{The outcome model .}
}
\value{
A tibble containing risks for exposure level 0 and 1, Risk Difference,
Risk ratio, Odds Ratio
}
\description{
A function to calculate effect of a binary exposure (i.e. treatment) on the
outcome using standardization in outcome-modeling approach.
}
